% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWESalyzer.R
\name{launch_GWESalyzer}
\alias{launch_GWESalyzer}
\title{Launch GWESalyzer}
\usage{
launch_GWESalyzer(launch_in_browser = TRUE)
}
\arguments{
\item{launch_in_browser}{Boolean value, true by default. If set to true, launches GWESalyzer in the user's default browser.}
}
\description{
Launch the Shiny app for GWESalyzer.
}
