% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWESExplorer.R
\name{launch_GWESExplorer}
\alias{launch_GWESExplorer}
\title{Launch GWES-Explorer}
\usage{
launch_GWESExplorer(launch_in_browser = TRUE)
}
\arguments{
\item{launch_in_browser}{Boolean value, true by default. If set to true, launches GWES-Explorer in the user's default browser.}
}
\description{
Launch the Shiny app for GWES-Explorer.
}
